Release 5.2

+++ os/services/lwm2m/lwm2m-json.h
@@ -49,8 +49,8 @@ struct json_data {
   uint8_t type; /* S,B,V */
   uint8_t *name;
   uint8_t *value;
-  uint8_t name_len;
-  uint8_t value_len;
+  uint16_t name_len;
+  uint16_t value_len;
 };
 
 extern const lwm2m_writer_t lwm2m_json_writer;
